Style update, refactor, and adding objects
This updates the project to the 2018 edition of Rust, and reformats the
code with rustfmt.

This also refactors the `sphere` module into an `object` module, and
creates the `Object` trait to act as a generic Ray Trace-able object
interface.

The `Material` struct is also added to hold material information, which
for now is just diffuse color.